requirements.txt
streamlit
xai-sdk
python-dotenv